feat: Added basic headers for proxy collector
This commit introduces basic headers to the proxy collector, which will help with the extraction process. The headers are aimed at mimicking a common user's browser and will be used in the network requests made by the collector. The specific headers included are:

- Accept
- Accept-Encoding
- Accept-Language
- Content-Type
- Sec-Ch-Ua-Platform
- Sec-Fetch-Mode
- Upgrade-Insecure-Requests
- User-Agent
- Referer

These headers will likely improve the success rate of proxy extraction and increase the overall reliability of the collector.